Add more error handling and recovery to preload()
An instance of a type stored in Vessel may be un-serializable when
the definition of that type changes.

This can occur when a type's definition is changed but an instance
of that type stored in Vessel is not removed or migrated.

The `preload()` and `preloadBlocking` APIs can now recover from this state
and continue to preload any remaining valid data.

`preload()`/`prealoadBlocking()` now return the result of the preload operation -
which errors ocurred (if any), and if the preload timeout was hit.

Since this is an API change, this is also a breaking change.

For completeness this same reporting has been added to the profiling data.